[refactor] : countBy 반환타입 int -> long으로 모두 변경

+++ survey/application/src/main/java/me/nalab/survey/application/service/findfeedbacksummary/FeedbackSummaryFindService.java
@@ -15,12 +15,12 @@ public class FeedbackSummaryFindService implements FeedbackSummaryFindUseCase {
 	private final UpdatedFeedbackCountPort updatedFeedbackCountPort;
 
 	@Override
-	public int getTotalFeedbackCount(Long surveyId) {
+	public long getTotalFeedbackCount(Long surveyId) {
 		return totalFeedbackCountPort.getTotalFeedbackCountBySurveyId(surveyId);
 	}
 
 	@Override
-	public int getUpdatedFeedbackCount(Long surveyId) {
+	public long getUpdatedFeedbackCount(Long surveyId) {
 		return updatedFeedbackCountPort.getUpdatedFeedbackCountBySurveyId(surveyId);
 	}
 }
